+++ tests/Populator/ContextMappingPopulatorTest.php
@@ -51,6 +51,19 @@ class ContextMappingPopulatorTest extends TestCase
         self::assertNull($person->getLocale());
     }
 
+    public function test_populateWithoutContextPropertyNames_required_throws(): void
+    {
+        $populator = new ContextMappingPopulator('locale', 'locale', null, null, null, true);
+        $user = new User();
+        $person = new Person();
+        $ctx = new GenericContext();
+
+        $this->expectException(PopulationException::class);
+        $this->expectExceptionMessage('The context does not contain a value for "locale"');
+
+        $populator->populate($person, $user, $ctx);
+    }
+
     public function test_populate_exceptional_case(): void
     {
         $populator = new ContextMappingPopulator('unknown', 'locale');
@@ -75,6 +88,18 @@ class ContextMappingPopulatorTest extends TestCase
         self::assertNull($person->getLocale());
     }
 
+    public function test_populateWithoutContext_required_throws(): void
+    {
+        $populator = new ContextMappingPopulator('locale', 'locale', null, null, null, true);
+        $user = new User();
+        $person = new Person();
+
+        $this->expectException(PopulationException::class);
+        $this->expectExceptionMessage('No context was provided.');
+
+        $populator->populate($person, $user);
+    }
+
     public function test_populateWithNewContext(): void
     {
         $populator = new ContextMappingPopulator('age', 'age', null, null, AgeContext::class);
@@ -97,6 +122,19 @@ class ContextMappingPopulatorTest extends TestCase
         $populator->populate($person, $user, $ctx);
 
         self::assertNull($person->getAge());
+    }
+
+    public function test_populateWithNewContext_missing_object_required_throws(): void
+    {
+        $populator = new ContextMappingPopulator('age', 'age', null, null, AgeContext::class, true);
+        $user = new User();
+        $person = new Person();
+        $ctx = Context::create();
+
+        $this->expectException(PopulationException::class);
+        $this->expectExceptionMessage(\sprintf('The context does not contain an instance of "%s"', AgeContext::class));
+
+        $populator->populate($person, $user, $ctx);
     }
 
     public function test_populateWithNewContext_missing_class_throws(): void